Update discovery.script
This script, although not idempotent, correctly assigns bolt preloads and bolt materials.  It assumes all other bodies are aluminium (which is correct for the components in the database).

It is very very slow and the material assignment is very bad, probably for the same reason: every bolt has a new material created for it.